warc_align_scoring/requirements.txt
beautifulsoup4==4.3.2
decorator==3.4.2
futures==3.0.2
goslate==1.4.0
htmllaundry==2.0
lxml==3.4.4
py==1.4.26
pytidylib==0.2.4
retry==0.9.0
six==1.9.0
